Fix IndexPath construction
The index paths in my example are all two-dimensional (section and
item), so we must construct them that way (assigning a section of 0 to
all works).

This fixes the crashes.

General observation: you're using way too many optionals when it's not
necessary. I fixed one instance, but they are many more. For example,
it's almost never a good idea to have an optional array of something
(as in `var indexPaths: [IndexPath]? = []`) because an empty array is
practically the same thing as no array at all.